requirements_txt.txt
streamlit>=1.28.0
openai>=1.12.0
tiktoken>=0.5.0
numpy>=1.24.0
pandas>=2.0.0
PyPDF2>=3.0.0
python-pptx>=0.6.21
pydrive2>=1.19.0
faiss-cpu>=1.7.4